Add a way to provide missing engine specificities for code generation
Try to import a dialect specific module to inject missing declaration
from SQLAlchemy (eg. Point and LTree types, or Postgis types).
This prevents missing type errors on generation but the project won't
work out of the box without a proper implementation of the type, which
should be provided as an external package like GeoAlchemy2.